corrected missing shellerror type
