+++ crates/nu-command/Cargo.toml
@@ -54,7 +54,7 @@ crossterm = "0.22.1"
 num = {version="0.4.0", optional=true}
 
 [dependencies.polars]
-version = "0.17.0"
+version = "0.18.0"
 optional = true
 features = ["default", "parquet", "json", "serde", "object", "checked_arithmetic", "strings"]
