+++ crates/nu-command/src/dataframe/README.md
@@ -1,3 +1,3 @@
 # nu-dataframe
 
-The nu-dataframe crate holds the definitions of the dataframe structure
+The nu-dataframe crate holds the definitions of the dataframe structures and commands
