+++ crates/nu-command/src/dataframe/mod.rs
@@ -1,3 +1,3 @@
 mod nu_dataframe;
 
-pub use nu_dataframe::commands::{DataTypes, DescribeDF, OpenDataFrame, ToDataFrame};
+pub use nu_dataframe::{DataTypes, DescribeDF, OpenDataFrame, ToDataFrame};
