--- crates/nu-command/src/dataframe/nu_dataframe/commands/mod.rs
@@ -1,12 +1,0 @@
-mod describe;
-mod dtypes;
-mod open;
-mod to_df;
-
-pub use describe::DescribeDF;
-pub use dtypes::DataTypes;
-pub use open::OpenDataFrame;
-pub use to_df::ToDataFrame;
-
-#[cfg(test)]
-mod test_dataframe;
